Replace dropbox-sdk files with dropbox-aiohttp files
